requirements.txt
aniso8601==9.0.1
click==8.0.0
et-xmlfile==1.1.0
Flask==2.0.0
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-OpenID==1.2.5
Flask-RESTful==0.3.9
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
phone==0.4.3
PyJWT==2.1.0
pymongo==3.11.4
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1
